refactor: Invalid class namespace
